Connecticut future preparation
In Connecticut, counties ceased most administrative functions in 1960, but remained for statistical purposes until being replaced by new "planning areas" with different boundaries in 2022. As all elections in the dataset took place prior to this change, data integrity won't be an issue for analysis. However, visualization tools such as Tableau may not know how to reconcile the old counties with the new planning areas. This will become a larger problem when/if I add 2024 data. I've added a notebook pre-emptively for such an event, in which I may have to use geospatial data at the precinct level to reconcile the differing boundaries.